Improve COM port filtering
Previously, this checked for the port description field, but it turns
out that Windows 10 does not use the description from the inf file, but
a more generic one.

This now looks up the hardware identifiers for the comport and checks
for the USB identifiers ("USB\VID_16D0&PID_0C5B").