SerialPortService: Change arrival and removal handling
Previously only serial devices were handled with this class, since downloads via
serial where the only use case, the rest was being dropped as arrival or removal
event.
For STM support however, firmware can be downloaded via USB composite devices
which doesn't show up as serial COM-port in OS's. The con of this is that every
composite interface and the base device events will be handled as separate
arrival and removal event. Meaning that if a device is plugged in and has 4
interfaces, the arrival event will trigger 5 times. To prevent an overkill of
events like that, we set up a wait timer which triggers after half a second
after receiving the first event and handles that after the timer elapses.