AvrCmdLine: Move avr specific commandline variables
The presets were first set in the 'Presets.cs', but the variables all existed in
the Form1 file. Those variables are not really related to the interface, so
these are probably better of in the AvrCmdLine class. Most variables are set
over there anyway. The con of this is only that the presets are not saved in the
 'presets.xml' file, but we won't be using those presets anyway at the moment
since the application should only upload to known 3devo products.