Executable: New implementation of Executable class
This implementation uses 2 threads, one for the stdout and one for stderr. Both
merge their output to the logger method which adds the text to a single string
object.
Although it hasn't been tested much, it does seem to get the job done for now.
Some methods might need some re-factoring later on.