HACK: Add as small delay to the 'ConsoleOutputHandler()'
Adding a small delay in the handler helps with displaying the output to the
infobox in the interface. When the handler is raised too quickly it doesn't
get triggered and therefore doesn't receive the data at all. In order to prevent
that from happening, even the slightest delay helps to solve this situation.
Perhaps this can be solved by delaying the calls if data is received (in the
executable class Logger() method?) within a certain time frame, but that would
also mean that the progress bar won't be as accurate in showing the duration
any more.