Form1: Fix machine information after device removal from USB
With the old structure the dropdown selectionChanged handler is not raised (and
therefore neither the 'ComboboxDropdown_Handler') when a device is removed, to
overcome this shortcoming we now call the handler manually.
Note that the ComboboxDropdown_Handler is also called when the selectionIndex is
changed manually, so there's no need to call the handler again after setting the
selectionIndex value.