HACK; Form1: Add a delay after the upload process is finished
USB-devices, using MCU's like ST with native USB, need to re-enumerate after the
upload process. Before the download process they will temporarily be reset to
the bootloader (as generic ST device) and when the download process is completed
they will reset again and continue to flash memory (and thus running the newly
flashed firmware). But all the resets also cause the application to re-evaluate
the devices connected to the computer. However the last reset to flash memory
could take a little while longer and cause the list to be refreshed too soon
(and sometimes also miss the last arrival event of the re-enumeration). Since
the GUI is already frozen (nothing is selectable), we might as well bridge the
enumeration time by simply delaying for a bit and manually get the latest device
information.