Revise DeviceList and introduce DevoHardware
This commit moves all 3devo hardware related ID's and information to
the DevoHardware class and implements this list usage in the DeviceList class.
It cleans up a lot of hardware specific code and makes the DeviceList class
more generic. This also simplifies the process of adding new devices, as we
only need to make changes to the DevoHardware class.
To add a device add a new device name to the "DEVO_HARDWARE_NAME" enum and add a
new device to the SerialDevices or DfuDevices (depending on which architecture
is used). Lastly add a user-friendly name to the GetString method if needed.
Serial devices already advertise their user friendly name, for Dfu devices
this may also be true but that query values couldn't be found easily.