Dfuse: Open files as readonly
Firmware uploads should be done with readonly permission to prevent
accidental writes to the selected file (especially for firmware). This also
helps when firmware is located in readonly folders, because selecting a file
without write permission would otherwise show a "no access to file x" error in
the update tool.